refactor(treewide): modularize weather methods
