+++ Cargo.toml
@@ -6,7 +6,9 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+tokio = { version = "1.27.0", features = ["full"] }
 actix-web = "4.3.0"
+anyhow = "1.0.70"
 futures = "0.3.26"
 lazy_static = "1.4.0"
 protobuf = "3.2.0"
@@ -17,7 +19,6 @@ sha2 = "0.10.6"
 subtle = "2.4.1"
 tracing = "0.1.37"
 tracing-subscriber = "0.3.17"
-#surrealdb = "1.0.0-beta.8"
 
 [build-dependencies]
 protobuf-codegen = "3.2.0"
