+++ src/geocoding/entities.rs
@@ -1,6 +1,6 @@
-use std::collections::HashMap;
-
+use crate::{weather::entities::ProtoAdapter, weather_proto::weather_message};
 use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub(crate) struct DoGeocodeResp {
@@ -10,4 +10,33 @@ pub(crate) struct DoGeocodeResp {
     pub(crate) lon: f64,
     pub(crate) country: String,
     pub(crate) state: Option<String>,
+}
+
+pub(crate) struct ReverseGeocode {
+    pub(crate) name: String,
+    pub(crate) country: String,
+    pub(crate) state: String,
+}
+
+impl ReverseGeocode {
+    pub(crate) fn default() -> Self {
+        Self {
+            name: "".to_string(),
+            country: "".to_string(),
+            state: "".to_string(),
+        }
+    }
+}
+
+impl ProtoAdapter for ReverseGeocode {
+    type ProtoType = weather_message::ReverseGeocode;
+
+    fn to_proto(&self) -> Self::ProtoType {
+        weather_message::ReverseGeocode {
+            name: self.name.to_owned(),
+            country: self.country.to_owned(),
+            state: self.state.to_owned(),
+            ..Default::default()
+        }
+    }
 }
