+++ src/geocoding/methods.rs
@@ -1,8 +1,8 @@
 use reqwest::StatusCode;
 
-use crate::{data_types, APIKey, Location};
+use crate::{APIKey, entities::Location};
 
-use super::entities::DoGeocodeResp;
+use super::entities::{DoGeocodeResp, ReverseGeocode};
 
 pub(crate) async fn do_geocode(keys: &APIKey, place_name: String) -> Location {
     if !keys.owm_key.is_empty() {
@@ -40,33 +40,42 @@ pub(crate) async fn do_geocode(keys: &APIKey, place_name: String) -> Location {
 pub(crate) async fn do_reverse_geocode(
     keys: &APIKey,
     location: &Location,
-) -> data_types::ReverseGeocode {
+) -> ReverseGeocode {
     if !keys.owm_key.is_empty() {
+        // construct query URL
         let owm_query = format!(
             "http://api.openweathermap.org/geo/1.0/reverse?lat={}&lon={}&limit=1&appid={}",
             location.latitude, location.longitude, keys.owm_key
         );
+
+        // make request
         let result = reqwest::get(owm_query).await;
+
+        // handle errors
         if let Ok(response) = result {
+            // Our request failed for some reason, we will try again later.
             if !StatusCode::is_success(&response.status()) {
-                // Our request failed for some reason, we will try again later.
-                return data_types::ReverseGeocode::default();
+                return ReverseGeocode::default();
             }
+
+            // deserialize response
             let response_mapping = response.json::<Vec<DoGeocodeResp>>().await.unwrap();
+
+            // return first response
             let first = response_mapping.get(0);
             if let Some(loc) = first {
-                data_types::ReverseGeocode {
+                ReverseGeocode {
                     name: loc.name.clone(),
                     country: loc.country.clone(),
                     state: loc.state.clone().unwrap_or(String::from("")),
                 }
             } else {
-                data_types::ReverseGeocode::default()
+                ReverseGeocode::default()
             }
         } else {
-            data_types::ReverseGeocode::default()
+            ReverseGeocode::default()
         }
     } else {
-        data_types::ReverseGeocode::default()
+        ReverseGeocode::default()
     }
 }
