+++ src/main.rs
@@ -1,62 +1,20 @@
 // mod database_utils;
-mod data_types;
+mod entities;
 mod geocoding;
+mod weather;
+mod utils;
 
-use crate::data_types::{Conditions, ProtoAdapter as _};
+use crate::weather::entities::{ProtoAdapter as _, Units};
 
-use crate::weather_proto::weather_message;
+use crate::weather::methods::do_weather_query;
 use actix_web::{get, web, App, HttpServer, Responder};
-use protobuf::Message;
-use reqwest::StatusCode;
+use entities::Location;
 use serde::Deserialize;
-use serde_json::Value;
 use tracing::info;
-use std::collections::HashMap;
 use std::env;
-use std::fmt::{Display, Formatter};
 
 mod weather_proto {
     include!(concat!(env!("OUT_DIR"), "/proto/mod.rs"));
-}
-
-struct Location {
-    latitude: f64,
-    longitude: f64,
-}
-
-impl Location {
-    fn default() -> Self {
-        Self {
-            latitude: 0.0,
-            longitude: 0.0,
-        }
-    }
-}
-
-enum Units {
-    Metric,
-    Imperial,
-    Standard,
-}
-
-impl Display for Units {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Units::Metric => f.write_str("metric"),
-            Units::Imperial => f.write_str("imperial"),
-            Units::Standard => f.write_str("standard"),
-        }
-    }
-}
-
-impl From<String> for Units {
-    fn from(unit: String) -> Self {
-        match unit.as_str() {
-            "metric" => Units::Metric,
-            "imperial" => Units::Imperial,
-            _ => Units::Standard,
-        }
-    }
 }
 
 #[derive(Deserialize, Debug)]
@@ -64,195 +22,9 @@ struct APIKey {
     owm_key: String,
 }
 
-fn convert_id_to_condition(current_weather_id: i64) -> Conditions {
-    if current_weather_id == 800 {
-        Conditions::Clear
-    } else if current_weather_id > 800 && current_weather_id < 805 {
-        if current_weather_id == 804 {
-            Conditions::Overcast
-        } else {
-            Conditions::Cloudy
-        }
-    } else if current_weather_id > 599 && current_weather_id < 700 {
-        Conditions::Snow
-    } else if current_weather_id > 199 && current_weather_id < 300 {
-        Conditions::Storm
-    } else {
-        Conditions::Rainy
-    }
-}
-
-fn process_current_weather(
-    current_weather_mapping: HashMap<String, Value>,
-) -> data_types::HourlyWeather {
-    let current_condition = {
-        // Reference: https://openweathermap.org/weather-conditions#Weather-Condition-Codes-2
-        let current_weather_weather: HashMap<String, Value> = serde_json::from_value(
-            current_weather_mapping
-                .get("weather")
-                .unwrap()
-                .as_array()
-                .unwrap()[0]
-                .clone(),
-        )
-        .unwrap();
-        let current_weather_id = current_weather_weather.get("id").unwrap().as_i64().unwrap();
-        convert_id_to_condition(current_weather_id)
-    };
-    data_types::HourlyWeather {
-        temp: current_weather_mapping
-            .get("temp")
-            .unwrap()
-            .as_f64()
-            .unwrap(),
-        feels_like: current_weather_mapping
-            .get("feels_like")
-            .unwrap()
-            .as_f64()
-            .unwrap(),
-        time: current_weather_mapping.get("dt").unwrap().as_i64().unwrap(),
-        condition: current_condition,
-    }
-}
-
-fn process_hourly_weather(
-    hourly_weather_mapping: Vec<HashMap<String, Value>>,
-) -> Vec<data_types::HourlyWeather> {
-    let mut result = vec![];
-    for hourly_weather in hourly_weather_mapping {
-        result.push(process_current_weather(hourly_weather));
-    }
-    result.sort_by(|e, e2| e.time.partial_cmp(&e2.time).unwrap());
-    result
-}
-
-fn process_daily_weather(
-    daily_weather_mapping: Vec<HashMap<String, Value>>,
-) -> Vec<data_types::OneDayForecast> {
-    let mut result = vec![];
-    for daily_weather in daily_weather_mapping {
-        println!("daily_weather: {:?}", daily_weather);
-        let temp_mapping: HashMap<String, f64> =
-            serde_json::from_value(daily_weather.get("temp").unwrap().clone()).unwrap();
-        let high_temp = *temp_mapping.get("max").unwrap();
-        let low_temp = *temp_mapping.get("min").unwrap();
-        let current_condition = {
-            // Reference: https://openweathermap.org/weather-conditions#Weather-Condition-Codes-2
-            let current_weather_weather: HashMap<String, Value> = serde_json::from_value(
-                daily_weather.get("weather").unwrap().as_array().unwrap()[0].clone(),
-            )
-            .unwrap();
-            let current_weather_id = current_weather_weather.get("id").unwrap().as_i64().unwrap();
-            convert_id_to_condition(current_weather_id)
-        };
-        result.push(data_types::OneDayForecast {
-            high_temp,
-            low_temp,
-            condition: current_condition,
-            time: daily_weather.get("dt").unwrap().as_i64().unwrap(),
-            sunrise: daily_weather.get("sunrise").unwrap().as_i64().unwrap(),
-            sunset: daily_weather.get("sunset").unwrap().as_i64().unwrap(),
-            rain: daily_weather
-                .get("rain")
-                .unwrap_or(&serde_json::Value::from(0f64))
-                .as_f64()
-                .unwrap(),
-        })
-    }
-    result
-}
-
-async fn do_aqi_query(keys: &APIKey, location: &Location) -> i32 {
-    if !keys.owm_key.is_empty() {
-        let owm_query = format!(
-            "http://api.openweathermap.org/data/2.5/air_pollution?lat={}&lon={}&appid={}",
-            location.latitude, location.longitude, keys.owm_key
-        );
-        let result = reqwest::get(owm_query).await;
-        if let Ok(response) = result {
-            // Our request failed for some reason, we will try again later.
-            let response_mapping: HashMap<String, Value> = response.json().await.unwrap();
-            let response_list: Vec<HashMap<String, Value>> =
-                serde_json::from_value(response_mapping.get("list").unwrap().clone()).unwrap();
-            let main: HashMap<String, i32> =
-                serde_json::from_value(response_list[0].get("main").unwrap().clone()).unwrap();
-            *main.get("aqi").unwrap()
-        } else {
-            -1
-        }
-    } else {
-        -1
-    }
-}
-
-fn convert_aqi_to_string(aqi: i32) -> String {
-    if aqi == 1 {
-        "Good".to_string()
-    } else if aqi == 2 {
-        "Fair".to_string()
-    } else if aqi == 3 {
-        "Moderate".to_string()
-    } else if aqi == 4 {
-        "Poor".to_string()
-    } else {
-        "Very poor".to_string()
-    }
-}
-
-async fn do_weather_query(keys: APIKey, location: Location, units: Units) -> impl Responder {
-    if !keys.owm_key.is_empty() {
-        let owm_query = format!(
-            "http://api.openweathermap.org/data/3.0/onecall?lat={}&lon={}&appid={}&units={}",
-            location.latitude, location.longitude, keys.owm_key, units
-        );
-        let result = reqwest::get(owm_query).await;
-        if result.is_err() {
-            println!("{:?}", result.err());
-            // Our request failed for some reason, we will try again later.
-            return b"request failed 1".to_vec();
-        }
-        let response = result.unwrap();
-        if !StatusCode::is_success(&response.status()) {
-            // Our request failed for some reason, we will try again later.
-            return format!("request failed with statuscode: {}", &response.status())
-                .as_bytes()
-                .to_vec();
-        }
-        let response_mapping: HashMap<String, Value> = response.json().await.unwrap();
-        eprintln!(
-            "current weather: {}",
-            response_mapping.get("current").unwrap()
-        );
-        let current_weather_mapping: HashMap<String, Value> = serde_json::from_str(
-            response_mapping
-                .get("current")
-                .unwrap()
-                .to_string()
-                .as_str(),
-        )
-        .unwrap();
-        let current_weather = process_current_weather(current_weather_mapping);
-        let hourly_weather_mapping =
-            serde_json::from_value(response_mapping.get("hourly").unwrap().clone()).unwrap();
-        let hourly_weather = process_hourly_weather(hourly_weather_mapping);
-        let daily_weather_mapping =
-            serde_json::from_value(response_mapping.get("daily").unwrap().clone()).unwrap();
-        let daily_weather = process_daily_weather(daily_weather_mapping);
-        let final_weather = weather_message::WeatherInfo {
-            hour_forecasts: hourly_weather.iter().map(|w| w.to_proto()).collect(),
-            current_weather: Some(current_weather.to_proto()).into(),
-            forecasts: daily_weather.iter().map(|w| w.to_proto()).collect(),
-            aqi: convert_aqi_to_string(do_aqi_query(&keys, &location).await),
-            geocode: Some(geocoding::methods::do_reverse_geocode(&keys, &location).await.to_proto()).into(),
-            ..Default::default()
-        };
-        return final_weather.write_to_bytes().unwrap();
-    }
-    b"no key".to_vec()
-}
-
 #[tracing::instrument]
 fn get_api_key_from_env() -> APIKey {
+    info!("Fetching OWM API key");
     // Load openweathermap api key and return it.
     let owm_key = env::var("OWM_KEY").expect("OWM_KEY not set");
     // kill the process if the key is empty and return the key if it is not.
