+++ src/weather/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod entities;
+pub(crate) mod methods;
+pub(crate) mod utils;
